Support for file versioning

+++ fields.py
@@ -3,6 +3,7 @@
 from osv.fields import char
 from tools import human_size
 
+import pymongo
 import gridfs as gfs
 from mongodb2 import mdbpool
 from bson.objectid import ObjectId
@@ -18,6 +19,7 @@ class gridfs(char):
     _type = 'char'
 
     def __init__(self, string, **args):
+        self.versioning = False
         super(gridfs, self).__init__(string=string, size=24, widget='binary',
                                      **args)
 
@@ -27,31 +29,49 @@ class gridfs(char):
         res = dict([(x[0], x[1]) for x in cursor.fetchall()])
         return res
 
+    def get_filename(self, obj, rid, name):
+        return '%s/%s_%s' % (obj._table, rid, name)
+
     def set(self, cursor, obj, rid, name, value, user=None, context=None):
         # TODO: Store some more metadata. File name, author, etc.
-        fs = gfs.GridFS(mdbpool.get_db(), collection='fs')
-        # Maybe support file versioning using obj._table + _rid?
+        db = mdbpool.get_db()
+        fs = gfs.GridFS(db, collection='fs')
         for rid, oid in self.get_oids(cursor, obj, [rid], name).items():
-            if oid and fs.exists(ObjectId(oid)):
+            filename = self.get_filename(obj, rid, name)
+            if oid and fs.exists(ObjectId(oid)) and not self.versioning:
                 fs.delete(ObjectId(oid))
             if value:
-                _id = fs.put(value)
+                _id = fs.put(value, filename=filename)
                 value = str(_id)
+            if not value and self.versioning:
+                fs.delete(ObjectId(oid))
+                res = db.fs.files.find(
+                    {'filename': filename},
+                    {'uploadDate': True, '_id': True}
+                ).sort('filename', pymongo.DESCENDING).limit(1)
+                if res.count():
+                    value = str(res[0]['_id'])
             return super(gridfs, self).set(cursor, obj, rid, name, value, user,
                                            context)
 
     def get(self, cursor, obj, ids, name, user=None, offset=0, context=None,
             values=None):
         if not context:
-            context= {}
-        fs = gfs.GridFS(mdbpool.get_db(), collection='fs')
+            context = {}
+        db = mdbpool.get_db()
+        fs = gfs.GridFS(db, collection='fs')
         res = self.get_oids(cursor, obj, ids, name)
         for rid, oid in res.items():
+            filename = self.get_filename(obj, rid, name)
             if oid:
                 oid = ObjectId(oid)
                 val = fs.get(oid).read()
                 if context.get('bin_size', False) and val:
-                    res[rid] = human_size(val)
+                    version = db.fs.files.find(
+                        {'filename': filename},
+                        {'uploadDate': True, '_id': False}
+                    ).count()
+                    res[rid] = '%s - v%s' % (human_size(val), version)
                 else:
                     res[rid] = val
             else:
